+++ frontend/components/screens/feed/Feed.tsx
@@ -22,6 +22,7 @@ import { useRouter } from 'next/router';
 import { feedTabMenus } from '@/components/ui/tab-menu/tab-menu.constants';
 import MyFeed from './my-feed/MyFeed';
 import { useFeedLike } from '@/hooks/useFeedLike';
+import { useCommentLike } from '@/hooks/useCommentLike';
 
 const Feed: FC = () => {
 	const router = useRouter();
@@ -76,42 +77,42 @@ const Feed: FC = () => {
 	// 	},
 	// );
 
-	const { mutate: commentLikeSync } = useMutation(
-		['comment-like'],
-		(data: { feedId: string; commentId: string; page: number }) =>
-			CommentService.updateLike(data.feedId, data.commentId),
-		{
-			onMutate: variable => {
-				//Loading.hourglass();
-			},
-			onSuccess(data, variable) {
-				const pageValue = variable.page;
-				if (data === true) {
-					handleIsLottie(true);
-					Notify.success('좋아요를 누르셨습니다');
-				}
-				if (data === false) {
-					Notify.warning('좋아요를 취소하셨습니다');
-				}
+	// const { mutate: commentLikeSync } = useMutation(
+	// 	['comment-like'],
+	// 	(data: { feedId: string; commentId: string; page: number }) =>
+	// 		CommentService.updateLike(data.feedId, data.commentId),
+	// 	{
+	// 		onMutate: variable => {
+	// 			//Loading.hourglass();
+	// 		},
+	// 		onSuccess(data, variable) {
+	// 			const pageValue = variable.page;
+	// 			if (data === true) {
+	// 				handleIsLottie(true);
+	// 				Notify.success('좋아요를 누르셨습니다');
+	// 			}
+	// 			if (data === false) {
+	// 				Notify.warning('좋아요를 취소하셨습니다');
+	// 			}
 
-				refetch({ refetchPage: (page, index) => index === pageValue - 1 });
+	// 			refetch({ refetchPage: (page, index) => index === pageValue - 1 });
 
-				//Loading.remove();
-				//if (data) Report.success('성공', `좋아요를 성공 하였습니다`, '확인');
-				//Report.success('성공', `좋아요를 취소 하였습니다`, '확인');
-			},
-			onError(error) {
-				if (axios.isAxiosError(error)) {
-					Report.warning(
-						'실패',
-						`${error.response?.data.message}`,
-						'확인',
-						() => Loading.remove(),
-					);
-				}
-			},
-		},
-	);
+	// 			//Loading.remove();
+	// 			//if (data) Report.success('성공', `좋아요를 성공 하였습니다`, '확인');
+	// 			//Report.success('성공', `좋아요를 취소 하였습니다`, '확인');
+	// 		},
+	// 		onError(error) {
+	// 			if (axios.isAxiosError(error)) {
+	// 				Report.warning(
+	// 					'실패',
+	// 					`${error.response?.data.message}`,
+	// 					'확인',
+	// 					() => Loading.remove(),
+	// 				);
+	// 			}
+	// 		},
+	// 	},
+	// );
 
 	const {
 		data,
@@ -187,13 +188,18 @@ const Feed: FC = () => {
 
 	const { handleUpdateLike } = useFeedLike({ handleRefetch, handleIsLottie });
 
-	const handleLikeComment = (
-		feedId: string,
-		commentId: string,
-		page: number,
-	) => {
-		commentLikeSync({ feedId, commentId, page });
-	};
+	const { handleLikeComment } = useCommentLike({
+		handleRefetch,
+		handleIsLottie,
+	});
+
+	// const handleLikeComment = (
+	// 	feedId: string,
+	// 	commentId: string,
+	// 	page: number,
+	// ) => {
+	// 	commentLikeSync({ feedId, commentId, page });
+	// };
 
 	return (
 		<Format title={'feed'}>
